requirement.txt
Django==1.8

Pillow==4.1.1
django-localflavor==1.5.2

django-widget-tweaks==1.4.1

djangorestframework==3.6.3

psycopg2==2.7.1

Pygments==2.2.0

pytz==2017.2

django-cors-header==2.1.0
numpy==1.12.1
olefile==0.44
setuptools==28.8.0
